COMP: Add CMAKE code to generate comprehensive DCMTKConfig files
Since CMake 2.6 (and in particular since 2.8) CMake supports a new signature for
find_package: find_package(<package_name> NO_MODULE)

NO_MODULE tells CMake to skip loading the Find<package>.cmake to locate a
package and use it. The big advantage for implementing this in a package like
DCMTK is that once it is properly implemented, a third-party library (like
DCMTK) can be used by other CMake projects without needing a Find<package>.cmake
file in the CMake installation.

This work in DCMTK is based on this tutorial:
http://www.cmake.org/Wiki/CMake/Tutorials/How_to_create_a_ProjectConfig.cmake_file

along with assistance from Brad King (brad.king@kitware.com)